+++ _Typings/gameApi/types/tween.d.ts
@@ -2,23 +2,23 @@
 // For details: https://github.com/Withered-Flower-0422/BST/blob/main/LICENSE
 
 declare module "game:type" {
-    type Tween = Readonly<{
+    type Tween = {
         /**
          * Plays the tween for one step.
          * @returns
          */
-        next: () => void
+        readonly next: () => void
 
         /**
          * Plays the tween until it is completed.
          * @returns
          */
-        play: () => void
+        readonly play: () => void
 
         /**
          * Stops the tween. The next call to `next()` or `play()` will restart the tween from the beginning.
          * @returns
          */
-        stop: () => void
-    }>
+        readonly stop: () => void
+    }
 }